requirements.txt
fastapi>=0.128.0,<0.129.0
uvicorn>=0.40.0,<0.41.0
psycopg2-binary>=2.9.11,<3.0.0
pandas>=2.3.3,<3.0.0
numpy>=1.26.0,<3.0.0
python-dotenv>=1.2.1,<2.0.0
fastapi-pagination>=0.15.6,<0.16.0
pydantic>=2.12.5,<3.0.0
python-multipart>=0.0.21,<0.0.22